Restore Digitiser and MZ for 1998-04-10.
There are two sets of Mega-Zine in the recording, but there are not enough frames of the earlier edition to make a good recovery. These pages are from the later set.

Digitiser 171/3... strong indication that the checksum is 0x2FDE, but I have looked over the page several times and can't see anything amiss.