Oracle's Farewell Message on P698
"ORACLE's service to you will close at midnight tonight. We were outbid in the Government's auction of teletext.
We are proud to have been part of the team that made teletext a success and brought news and entertainment to so many.
The entire ORACLE staff thank you for your support over the years."

Followed by a list of names.

(Two variations, Merry Christmas and Happy New Year. CRCs perfect.)